* Migrate CaloCellGeometry to DataFormats * Update Geometry packages to use it from new package
