+++ Geometry/CaloEventSetup/plugins/CaloTowerConstituentsMapBuilder.cc
@@ -21,7 +21,7 @@
 #include "Geometry/CaloTopology/interface/CaloTowerTopology.h"
 #include "Geometry/Records/interface/CaloGeometryRecord.h"
 #include "Geometry/CaloGeometry/interface/CaloSubdetectorGeometry.h"
-#include "Geometry/CaloGeometry/interface/CaloCellGeometry.h"
+#include "DataFormats/CaloGeometry/interface/CaloCellGeometry.h"
 #include "DataFormats/EcalDetId/interface/EEDetId.h"
 #include "DataFormats/HcalDetId/interface/HcalDetId.h"
 #include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
