+++ Geometry/CaloGeometry/interface/EcalTrapezoidParameters.h
@@ -1,4 +1,4 @@
-#if !defined(ECALCOMMONDATA_ECALTRAPEZOIDPARAMETERS_H)
+#ifndef ECALCOMMONDATA_ECALTRAPEZOIDPARAMETERS_H
 #define ECALCOMMONDATA_ECALTRAPEZOIDPARAMETERS_H
 // -*- C++ -*-
 //
@@ -53,107 +53,77 @@
 //
 // Author:      Brian K. Heltsley
 // Created:     Wed Aug 12 09:25:08 EDT 1998
-// system include files
-
-// user include files
 
 #include <vector>
 #include <CLHEP/Geometry/Point3D.h>
 
-#include "Geometry/CaloGeometry/interface/CaloCellGeometry.h"
-
-// forward declarations
-
+#include "DataFormats/CaloGeometry/interface/CaloCellGeometry.h"
 
 class EcalTrapezoidParameters
 {
-      // ---------- friend classes and functions ---------------
-   public:
-      // ---------- constants, enums and typedefs --------------
+ public:
 
-      typedef CaloCellGeometry::Pt3DVec VertexList ;
+  typedef CaloCellGeometry::Pt3DVec VertexList ;
+  typedef CaloCellGeometry::CCGFloat TPFloat ;
 
-      typedef CaloCellGeometry::CCGFloat TPFloat ;
+  EcalTrapezoidParameters( TPFloat aHalfLengthXNegZLoY , // bl1, A/2
+			   TPFloat aHalfLengthXPosZLoY , // bl2, a/2
+			   TPFloat aHalfLengthXPosZHiY , // tl2, b/2
+			   TPFloat aHalfLengthYNegZ    , // h1,  H/2
+			   TPFloat aHalfLengthYPosZ    , // h2,  h/2
+			   TPFloat aHalfLengthZ        , // dz,  L/2
+			   TPFloat aAngleAD            , // alfa1
+			   TPFloat aCoord15X           , // x15
+			   TPFloat aCoord15Y             // y15
+			   ) ;
 
-      // ---------- Constructors and destructor ----------------
+  // GEANT parameters, in order
+  TPFloat dz()    const ;
+  TPFloat theta() const ;
+  TPFloat phi()   const ;
+  TPFloat h1()    const ;
+  TPFloat bl1()   const ;
+  TPFloat tl1()   const ;
+  TPFloat alp1()  const ;
+  TPFloat h2()    const ;
+  TPFloat bl2()   const ;
+  TPFloat tl2()   const ;
+  TPFloat alp2()  const ;
 
-      EcalTrapezoidParameters( TPFloat aHalfLengthXNegZLoY , // bl1, A/2
-			       TPFloat aHalfLengthXPosZLoY , // bl2, a/2
-			       TPFloat aHalfLengthXPosZHiY , // tl2, b/2
-			       TPFloat aHalfLengthYNegZ    , // h1,  H/2
-			       TPFloat aHalfLengthYPosZ    , // h2,  h/2
-			       TPFloat aHalfLengthZ        , // dz,  L/2
-			       TPFloat aAngleAD            , // alfa1
-			       TPFloat aCoord15X           , // x15
-			       TPFloat aCoord15Y             // y15
-	 ) ;
+  // everything else
+  TPFloat x15()   const ;
+  TPFloat y15()   const ;
+  TPFloat hAa()   const ;
+  TPFloat hBb()   const ;
+  TPFloat hCc()   const ;
+  TPFloat hDd()   const ;
+  TPFloat a1()    const ;
+  TPFloat a4()    const ;
 
-      //virtual ~EcalTrapezoidParameters() ;
+  TPFloat L()     const ;
+  TPFloat a()     const ;
+  TPFloat b()     const ;
+  TPFloat c()     const ;
+  TPFloat d()     const ;
+  TPFloat h()     const ;
+  TPFloat A()     const ;
+  TPFloat B()     const ;
+  TPFloat C()     const ;
+  TPFloat D()     const ;
+  TPFloat H()     const ;
 
-      // ---------- member functions ---------------------------
-      // ---------- const member functions ---------------------
+  VertexList vertexList() const ; // order is as in picture above: index=vtx-1
 
-      // GEANT parameters, in order
-      TPFloat dz()    const ;
-      TPFloat theta() const ;
-      TPFloat phi()   const ;
-      TPFloat h1()    const ;
-      TPFloat bl1()   const ;
-      TPFloat tl1()   const ;
-      TPFloat alp1()  const ;
-      TPFloat h2()    const ;
-      TPFloat bl2()   const ;
-      TPFloat tl2()   const ;
-      TPFloat alp2()  const ;
+ private:
+ 
+  EcalTrapezoidParameters() = delete;
+  EcalTrapezoidParameters( const EcalTrapezoidParameters& ) = delete;
+  const EcalTrapezoidParameters& operator=( const EcalTrapezoidParameters& ) = delete;
 
-      // everything else
-      TPFloat x15()   const ;
-      TPFloat y15()   const ;
-      TPFloat hAa()   const ;
-      TPFloat hBb()   const ;
-      TPFloat hCc()   const ;
-      TPFloat hDd()   const ;
-      TPFloat a1()    const ;
-      TPFloat a4()    const ;
-
-      TPFloat L()     const ;
-      TPFloat a()     const ;
-      TPFloat b()     const ;
-      TPFloat c()     const ;
-      TPFloat d()     const ;
-      TPFloat h()     const ;
-      TPFloat A()     const ;
-      TPFloat B()     const ;
-      TPFloat C()     const ;
-      TPFloat D()     const ;
-      TPFloat H()     const ;
-
-      VertexList vertexList() const ; // order is as in picture above: index=vtx-1
-
-      // ---------- static member functions --------------------
-   protected:
-      // ---------- protected member functions -----------------
-      // ---------- protected const member functions -----------
-   private:
-      // ---------- Constructors and destructor ----------------
-      EcalTrapezoidParameters();
-      EcalTrapezoidParameters( const EcalTrapezoidParameters& ); // stop default
-
-      // ---------- assignment operator(s) ---------------------
-      const EcalTrapezoidParameters& operator=( const EcalTrapezoidParameters& ); // stop default
-
-      // ---------- private member functions -------------------
-      // ---------- private const member functions -------------
-      // ---------- data members -------------------------------
-
-      TPFloat m_dz, m_th, m_ph, m_h1, m_bl1, m_tl1, m_alp1, m_h2, m_bl2, m_tl2, m_alp2 ;
-      TPFloat m_a1, m_hAa, m_x15, m_y15 ;
-      TPFloat m_a4, m_hBb, m_hCc, m_hDd ;
-      TPFloat m_L, m_a, m_b, m_c, m_d, m_h, m_A, m_B, m_C, m_D, m_H ;
-
-      // ---------- static data members ------------------------
+  TPFloat m_dz, m_th, m_ph, m_h1, m_bl1, m_tl1, m_alp1, m_h2, m_bl2, m_tl2, m_alp2 ;
+  TPFloat m_a1, m_hAa, m_x15, m_y15 ;
+  TPFloat m_a4, m_hBb, m_hCc, m_hDd ;
+  TPFloat m_L, m_a, m_b, m_c, m_d, m_h, m_A, m_B, m_C, m_D, m_H ;
 };
-
-// inline function definitions
 
 #endif /* ECALCOMMONDATA_ECALTRAPEZOIDPARAMETERS_H */
