+++ Geometry/EcalAlgo/BuildFile.xml
@@ -2,6 +2,7 @@
   <lib   name="1"/>
 </export>
 <use   name="Geometry/CaloGeometry"/>
+<use   name="DataFormats/CaloGeometry"/>
 <use   name="DataFormats/EcalDetId"/>
 <use   name="clhep"/>
   <use name="CondCore/DBOutputService"/>
