+++ Geometry/HcalTowerAlgo/src/HcalFlexiHardcodeGeometryLoader.cc
@@ -1,5 +1,4 @@
 #include "Geometry/HcalTowerAlgo/interface/HcalFlexiHardcodeGeometryLoader.h"
-#include "Geometry/CaloGeometry/interface/CaloCellGeometry.h"
 #include "Geometry/CaloGeometry/interface/IdealObliquePrism.h"
 #include "Geometry/CaloGeometry/interface/IdealZPrism.h"
 #include "Geometry/CaloTopology/interface/HcalTopology.h"
